% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagshelp.R
\name{bagsHelp}
\alias{bagsHelp}
\title{Show available commands in the bagsInstaller package}
\usage{
bagsHelp()
}
\value{
A printed help menu listing all exported functions.
}
\description{
Show available commands in the bagsInstaller package
}
