% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installer.R
\name{install_bags_dependencies}
\alias{install_bags_dependencies}
\title{Install all required dependencies for the BAGS app}
\usage{
install_bags_dependencies()
}
\description{
This function checks for missing packages and installs them.
}
